requirements.txt
# Postgres client
psycopg2-binary==2.9.5
# Curl client
requests==2.31.0
# Timezone library
pytz==2022.7.1
# MSSQL client
python-tds==1.12.0
# UI library
selenium==4.8.2
# SSO client
python-keycloak-client==0.2.3
# SFTP client
pysftp==0.2.9
# playwright
playwright==1.31.1
